requirements.txt
Flask==3.1.0
gunicorn==23.0.0
flask_sqlalchemy==3.1.1
flask_cors==5.0.1
pyjwt==2.10.1
bcrypt==4.3.0
